feat: 스탯 시스템 적용
unit -> entity 이름 변경, 스탯 시스템 적용

+++ TeamProject_TextRPG/ModifierSystem/StatMediator.cs
@@ -51,9 +51,9 @@ namespace TeamProject_TextRPG.ModifierSystem
     public class Query
     {
         public readonly StatType StatType;
-        public int Value;
+        public float Value;
 
-        public Query(StatType statType, int value)
+        public Query(StatType statType, float value)
         {
             StatType = statType;
             Value = value;
